+++ typespec-tests/src/main/java/com/cadl/partialupdate/models/PartialUpdateModel.java
@@ -45,10 +45,11 @@ public final class PartialUpdateModel implements JsonSerializable<PartialUpdateM
     @Generated
     private String aggregate;
 
+    // @formatter:off
     /**
      * customized property
      */
-    private String testProperty;
+    private String testProperty;// @formatter:on
 
     /**
      * Creates an instance of PartialUpdateModel class.
@@ -107,6 +108,7 @@ public final class PartialUpdateModel implements JsonSerializable<PartialUpdateM
         return this.aggregate;
     }
 
+    // @formatter:off
     /**
      * javadoc
      *
@@ -114,8 +116,9 @@ public final class PartialUpdateModel implements JsonSerializable<PartialUpdateM
      */
     public void setAggregate(String aggregate) {
         this.aggregate = aggregate;
-    }
+    }// @formatter:on
 
+    // @formatter:off
     /**
      * javadoc
      *
@@ -123,8 +126,9 @@ public final class PartialUpdateModel implements JsonSerializable<PartialUpdateM
      */
     public String getTestProperty() {
         return testProperty;
-    }
+    }// @formatter:on
 
+    // @formatter:off
     /**
      * javadoc
      *
@@ -132,8 +136,9 @@ public final class PartialUpdateModel implements JsonSerializable<PartialUpdateM
      */
     public void setTestProperty(String testProperty) {
         this.testProperty = testProperty;
-    }
+    }// @formatter:on
 
+    // @formatter:off
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
@@ -142,8 +147,9 @@ public final class PartialUpdateModel implements JsonSerializable<PartialUpdateM
         jsonWriter.writeBinaryField("bytes", this.bytes);
         jsonWriter.writeStringField("aggregate", this.aggregate);
         return jsonWriter.writeEndObject();
-    }
+    }// @formatter:on
 
+    // @formatter:off
     /**
      * Reads an instance of PartialUpdateModel from the JsonReader.
      *
@@ -178,5 +184,6 @@ public final class PartialUpdateModel implements JsonSerializable<PartialUpdateM
             deserializedPartialUpdateModel.aggregate = aggregate;
             return deserializedPartialUpdateModel;
         });
-    }
+    }// @formatter:on
+
 }