+++ typespec-extension/changelog.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 0.15.9 (2024-04-15)
+
+Compatible with compiler 0.55.
+
+- Update dependency of "typespec-client-generator-core" for bug fix.
+
 ## 0.15.8 (2024-04-11)
 
 Compatible with compiler 0.55.